backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2
email-validator==2.2.0

